Refactoring DeleUserUseCase and linking DeleteUserController to respective end-point

+++ src/modules/users/useCases/deleteUser/DeleteUserUseCase.ts
@@ -15,6 +15,12 @@ class DeleteUserUseCase {
   ) {}
 
   async execute({ id }: IRequest): Promise<User[]> {
+    const user = await this.usersRepository.findById(id)
+
+    if (!user) {
+      throw new Error('User does not exist')
+    }
+
     const users = await this.usersRepository.deleteUser(id)
 
     return users
